fix tests

+++ server/QuizBuilder.Test/Integration/WorkflowTests/Workflow2.cs
@@ -26,7 +26,7 @@ namespace QuizBuilder.Test.Integration.WorkflowTests {
 		public async Task Test() {
 
 			// Create Quiz 1
-			(HttpStatusCode statusCode, QuizCommandResult data) result1 = await _httpClient.PostValueAsync<QuizCommandResult>( "/quizzes/", new { Name = "Quiz 1" } );
+			(HttpStatusCode statusCode, QuizCommandResult data) result1 = await _httpClient.PostValueAsync<QuizCommandResult>( "admin/quizzes/", new { Name = "Quiz 1" } );
 			string uid1 = result1.data.Quiz.Id;
 
 			// Create Question 1
@@ -38,7 +38,7 @@ namespace QuizBuilder.Test.Integration.WorkflowTests {
 				Settings = "{\"choicesDisplayType\":1,\"choicesEnumerationType\":2}",
 				Choices = "[{\"isCorrect\":true,\"text\":\"Choice 1\"},{\"isCorrect\":false,\"text\":\"Choice 2\"}]"
 			};
-			(HttpStatusCode statusCode, QuestionCommandResult data) result2 = await _httpClient.PostValueAsync<QuestionCommandResult>( "/questions/", content1 );
+			(HttpStatusCode statusCode, QuestionCommandResult data) result2 = await _httpClient.PostValueAsync<QuestionCommandResult>( "admin/questions/", content1 );
 			string questionUId1 = result2.data.Question.Id;
 
 			// Create Question 2
@@ -53,11 +53,11 @@ namespace QuizBuilder.Test.Integration.WorkflowTests {
 				Settings = "{\"choicesDisplayType\":1,\"choicesEnumerationType\":2}",
 				Choices = "[{\"isCorrect\":true,\"text\":\"Choice 1\"},{\"isCorrect\":false,\"text\":\"Choice 2\"}]"
 			};
-			(HttpStatusCode statusCode, QuestionCommandResult data) result3 = await _httpClient.PostValueAsync<QuestionCommandResult>( "/questions/", content2 );
+			(HttpStatusCode statusCode, QuestionCommandResult data) result3 = await _httpClient.PostValueAsync<QuestionCommandResult>( "admin/questions/", content2 );
 			string questionUId2 = result3.data.Question.Id;
 
 			// Get All Questions
-			(HttpStatusCode statusCode, QuestionsQueryResult data) result4 = await _httpClient.GetValueAsync<QuestionsQueryResult>( $"/quizzes/{uid1}/questions/" );
+			(HttpStatusCode statusCode, QuestionsQueryResult data) result4 = await _httpClient.GetValueAsync<QuestionsQueryResult>( $"admin/quizzes/{uid1}/questions/" );
 			Assert.Equal( 2, result4.data.Questions.Count );
 			Assert.Equal( questionUId1, result4.data.Questions[0].Id );
 			Assert.Equal( questionUId2, result4.data.Questions[1].Id );
@@ -76,22 +76,22 @@ namespace QuizBuilder.Test.Integration.WorkflowTests {
 				Settings = "{\"choicesDisplayType\":1,\"choicesEnumerationType\":2}",
 				Choices = "[{\"isCorrect\":true,\"text\":\"Choice 1\"},{\"isCorrect\":false,\"text\":\"Choice 2\"}]"
 			};
-			await _httpClient.PutValueAsync<QuestionCommandResult>( "/questions/", content3 );
+			await _httpClient.PutValueAsync<QuestionCommandResult>( "admin/questions/", content3 );
 
 			// Get Question 2
-			(HttpStatusCode statusCode, QuestionQueryResult data) result6 = await _httpClient.GetValueAsync<QuestionQueryResult>( $"/questions/{questionUId2}" );
+			(HttpStatusCode statusCode, QuestionQueryResult data) result6 = await _httpClient.GetValueAsync<QuestionQueryResult>( $"admin/questions/{questionUId2}" );
 			Assert.Equal( "Question Name 2 New", result6.data.Question.Name );
 
 			// Delete Question 1
-			var result7 = await _httpClient.DeleteAsync( $"/quizzes/{uid1}/questions/{questionUId1}" );
+			var result7 = await _httpClient.DeleteAsync( $"admin/quizzes/{uid1}/questions/{questionUId1}" );
 			Assert.Equal( HttpStatusCode.NoContent, result7.StatusCode );
 
 			// Get Question 1
-			(HttpStatusCode statusCode, QuestionQueryResult data) result8 = await _httpClient.GetValueAsync<QuestionQueryResult>( $"/questions/{questionUId1}" );
+			(HttpStatusCode statusCode, QuestionQueryResult data) result8 = await _httpClient.GetValueAsync<QuestionQueryResult>( $"admin/questions/{questionUId1}" );
 			Assert.Equal( HttpStatusCode.NoContent, result8.statusCode );
 
 			// Get Question 2
-			(HttpStatusCode statusCode, QuestionQueryResult data) result9 = await _httpClient.GetValueAsync<QuestionQueryResult>( $"/questions/{questionUId2}" );
+			(HttpStatusCode statusCode, QuestionQueryResult data) result9 = await _httpClient.GetValueAsync<QuestionQueryResult>( $"admin/questions/{questionUId2}" );
 			Assert.Equal( "Question Name 2 New", result9.data.Question.Name );
 		}
